+++ config/upgrade.go
@@ -1,5 +1,5 @@
 // mautrix-whatsapp - A Matrix-WhatsApp puppeting bridge.
-// Copyright (C) 2021 Tulir Asokan
+// Copyright (C) 2022 Tulir Asokan
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU Affero General Public License as published by
@@ -40,6 +40,8 @@ func (helper *UpgradeHelper) doUpgrade() {
 	helper.Copy(Str, "appservice", "database", "uri")
 	helper.Copy(Int, "appservice", "database", "max_open_conns")
 	helper.Copy(Int, "appservice", "database", "max_idle_conns")
+	helper.Copy(Str|Null, "appservice", "database", "max_conn_idle_time")
+	helper.Copy(Str|Null, "appservice", "database", "max_conn_lifetime")
 	helper.Copy(Str, "appservice", "provisioning", "prefix")
 	if secret, ok := helper.Get(Str, "appservice", "provisioning", "shared_secret"); !ok || secret == "generate" {
 		sharedSecret := appservice.RandomString(64)